requirements.txt
customtkinter
pillow
requests
pyodbc
